In some cases the GC is still failing with the following error:

```
System.OverflowException: Arithmetic operation resulted in an overflow.
   at BuildXL.Cache.ContentStore.Distributed.NuCache.ContentLocationDatabase.GarbageCollectContentCore(OperationContext context) in \.\Public\Src\Cache\ContentStore\Distributed\NuCache\ContentLocationDatabase.cs:line 373
   at BuildXL.Cache.ContentStore.Distributed.NuCache.ContentLocationDatabase.<>c__DisplayClass36_0.<GarbageCollectContent>b__0() in \.\Public\Src\Cache\ContentStore\Distributed\NuCache\ContentLocationDatabase.cs:line 363
   at BuildXL.Cache.ContentStore.Tracing.Internal.PerformOperationBuilderBase`2.RunOperationAndConvertExceptionToError[T](Func`1 operation) in \.\Public\Src\Cache\ContentStore\Library\Tracing\OperationContextExtensions.cs:line 214
```

+++ Public/Src/Cache/ContentStore/Distributed/NuCache/ContentLocationDatabase.cs
@@ -375,12 +375,12 @@ namespace BuildXL.Cache.ContentStore.Distributed.NuCache
         private BoolResult GarbageCollectContentCore(OperationContext context)
         {
             // Counters for work done.
-            int removedEntries = 0;
-            int totalEntries = 0;
+            long removedEntries = 0;
+            long totalEntries = 0;
             long uniqueContentSize = 0;
             long totalContentCount = 0;
             long totalContentSize = 0;
-            int uniqueContentCount = 0;
+            long uniqueContentCount = 0;
 
             // Tracking the difference between sequence of hashes for diagnostic purposes. We need to know how good short hashes are and how close are we to collisions. 
             ShortHash? lastHash = null;
